Specify metric leader event and period/frequency
Use the option `--metric-leader <event-name>` to specify an event that
acts as a leader event when reading metrics given by `-E <event-name>`.
If `--metric-count n` is given, the metric buffer is read every `n`
times that the leader event has occured.  If `--metric-frequency f` is
given, the kernel is asked to read events with an approximate frequency
of `f` Hz.  The last two options are mutually exclusive.