Add perf::EventCollection, parse events only once
Store all requested metric events in a singleton.  Using this,
trace::Counters::metric_class does not need to repeat the logic of
parsing event names and adding default events.

This code also checks if the metric leader event was given again as
regular metric event and if so, ignores it (after giving notice).