🐛 Workaround for broken static linking
libotf2 need libm, but when trying to link MOSTLY (or ALL) static, this leads to a glibc bug, where libm misses some references. Therefor, we circumvent the problem by removing the statically linked libm.a and just let the g++ compiler link against libm.so, as it does so anyways... for now. THIS WILL FUCKING EXPLODE SOME DAY. Have fun future me.