Parsing tracepoint events: Inspecto errno-rum!
C++ iostreams are in no way designed to yield anything that resembles
meaningful errors.  Event less are they designed to throw exceptions.
So in order to generate meaningful errors should we be unable to parse
a tracepoint event, we use C file I/O.

Parsing a tracepoint event most likely fails when opening its ID file.
We check errno and issue a warning to the user which is hopefully more
helpful than errno itself.

EACCES:
    lo2s is probably missing permissions to access /sys/kernel/debug,
    which is mounted root-only by default.

ENOENT:
    The requested tracepoint event might not exist.

In case of failure, throw Even::ParseError, catch appropiately up the
stack. If found to be a unrecoverable error, log and die with a fatal
message.