Fixes #52 Improves I/O error messages
* Handle failure to open tracepoints more gracefully

Should I/O operations fail when parsing a tracepoint event from
/sys/kernel/debug, throw a tracepoint::EventFormat::Error.

The tracepoint::SwitchWriter requires sched/sched_switch to work, so if
that fails to open warn that /sys/kernel/debug might be read-only.  Then
throw std::system_error and let main() handle it.

The same goes for tracepoint::ExitReader and sched/sched_process_exit.

* Parsing tracepoint events: Inspecto errno-rum!

C++ iostreams are in no way designed to yield anything that resembles
meaningful errors.  Event less are they designed to throw exceptions.
So in order to generate meaningful errors should we be unable to parse
a tracepoint event, we use C file I/O.

Parsing a tracepoint event most likely fails when opening its ID file.
We check errno and issue a warning to the user which is hopefully more
helpful than errno itself.

EACCES:
    lo2s is probably missing permissions to access /sys/kernel/debug,
    which is mounted root-only by default.

ENOENT:
    The requested tracepoint event might not exist.

In case of failure, throw Even::ParseError, catch appropiately up the
stack. If found to be a unrecoverable error, log and die with a fatal
message.

* Add note about get_sched_*_event throwing in ctors

get_sched_{switch,process_exit}_event() throw if they fail to parse
their respective tracepoint events.  This is not recoverable, but we can
catch the exception in the constructor (using a function-try-block), log
the error and rethrow a meaningful exception.

* Check errno even when parsing with std::ifstreams

Calling open() on an std::ifstream behaves like calling open() on its
rdbuf, which in turn should (TM) be equivalent to calling std::fopen(),
which sets errno on error.