Improve the way we assign names to traced tasks
From now on, we obtain names for both traced processes and threads
(collectively referred to as 'tasks') from their comm value.  In order
to avoid excessive reads to the /proc filesystem, keeping track of
tasks' names now works like this:

* before starting the trace, get names for all running tasks using
    get_comms_for_running_processes
* if a ptrace event indicates that a new task was spawned, assign an
    initial name by reading the new task's comm value using
    get_{process,task}_comm
* if a ptrace events indicates that a task is about to exit, do *NOT*
    update its name again: either this task never changed its name and
    the initial name is still correct, or a change was reported as a
    PERF_RECORD_COMM event
* before finishing the trace, update task names yet again via
    get_comms_for_running_processes

NOTE: the last step might not be needed if we can make sure that we
never miss the creation of a task while tracing.