Rewrite of the 'exit_reader.hpp' using the PERF_RECORD_COMM event. (#106)
Rewrite of the 'exit_reader.hpp' using the PERF_RECORD_COMM event. This uses PERF_RECORD_COMM, as this is the only point where you can get the process names from inside of perf. Instead of requiring access to the `debugfs`, system monitoring with a perf_event_paranoid setting less or equal to zero is possible now.