fix(metrics): Use .scale and .unit for PMU events
This fixes #208

PMU events may come with a separate scaling factor, which needs to be
multiplied with the event read-out. This is located in [event_path].scale.

As I was touching the PMU event read code anyways I also added support
for parsing the [event_path].unit, which tells us the unit of the PMU event.