avoid potential race condition during construction
The ScopeMonitor constructor called start() on itself in the
constructor. start() deferred to the ThreadedMonitor::start(), where a
thread using the current object is created (here: ScopeMonitor).

As this happens at the end of the ScopeMonitor constructor, the thread
may now execute operations on the object *before* the constructor
finished. This renders calls to virtual functions to produce undefined
behavior.

This is fixed by moving the start() outside the constructor:
ScopeMonitor now only *prepares* the measurement thread, the caller has
to invoke start() themselves -- after construction finished. Existing
invocations are patched.

We encountered this using a very fine-tuned testing setup using
configuration-specific ulimits for open files, here 700. A higher limit
did not cause the race condition, where a lower ulimit displayed a
different error.