clean up existing monitors on error
The construction of a ScopeMonitor for every CPU may fail. If either
(1) a monitor is not constructed successfully or
(2) a monitor throws during start()
all already (successfully) constructed & started monitors are stopped
before the encountered exception is rethrown.

The old behavior occured in our test setups, and the improper cleanup
created an error message. This caused lo2s to crash on the improper
cleanup -- *before* the actual root cause was reported to the user.