trace registry setup: fix lifetime of variables
The old version used `auto&` as type for function results, and did not
handle the underlying lifetime of the objects correctly resulting in
undefined behavior. This is replaced by a copy (`auto`) where
modification of the variable is required, and `const auto&` where it is
not.

This resulted in the map of system_tree_node by package to not be
initialized fully (second package was missing), from where later lookups
to this map failed. This commit fixes this initialization and
eliminates the previously occuring lookup error from `map::at()`.