fix man page install
The man page build and install process now follows more sane defaults:
If the requirements for building the man pages (pod2man, gzip) are
available, it is built by default. Otherwise, man pages are disabled.
This man page support is reported to the user as an enabled or
disabled feature as part of the feature summary.

If the man page is enabled explicitly (`-DINSTALL_MANPAGE=ON`), but the
dependencies are not found the configuration is aborted. This breaks
previous behavior, which skipped man page generation with a warning and
carried on -- even with man pages explicitly enabled.

Additionally, the man page is now compressed with gzip before being
installed. This follows the requirement of the debian package maintainer
manual. In order to avoid potential conflict on multiple invocations,
gzip is invoked with --keep --force to *always* produce output and
*never* destroy the input.

The man page dependencies (pod2man, gzip) are explicitly mentioned in
the README.