fix(syscall): Replace sh script with libaudit
The sh script parsed the unistd.h header to get the mapping of names to
syscall nrs. This sh script however fails on headers with heavy amounts
of #ifdefs, which is the case for the aarch64 platform.

This replaces that sh script with libaudit, which has specific functions
for the lookup of syscall names.

Introducing this as an additional dependency should be fine, as it is a
dependency of systemd and thus should be already installed on most
systems

Fix off-by-one error

Address requested changes