add logo
Add logo, originally designed by Thomas Ilsche. The source files and
rendered versions are added in `share/logo`, including additional notes.
Inkscape is required to render and edit the logos, but for normal
operation using the rendered PNGs also uplaoded is more portable.
There are three flavors of the logo, transparent, with background, and
with font.