logo: rm links, render more formats
The linking between logo files (circle only saved once and linked to
other files) is removed. This linking enforces usage of pixel images in
inkscape, even when exporting to formats where this is not necessary
(e.g. PDFs). Now the circle logo is copied to *all* files. When
adjusting the logo make sure to update the other files.

More formats are now renderd into `share/logo/rendered/`, including PNG
and PDFs. Furthermore, the inkscape-SVGs in `share/logo/src/` are
rendered to plain SVGs: When editing, use the inkscape SVGs.

During testing it seems that the masking features in the exported plain
SVGs is not supported by all renderers. When in doubt, use PNGs or PDFs.