fix(metrics): Prevent NaNs in sampling from occuring
A code path in the metric readout code could lead to the generation of
a metric value of "NaN" (see counter_buffer.hpp explanatory
comment for more details).

As metrics are recorded in an accumulated fashion, a NaN metric value
will break the metric for the rest of the recording.

This commit deletes that code path, as there is no record of the bug it
tries to fix ever existing.

This commit further improves the resilence of the metric readout code by
introducing further checks for values of diff_running/enabled/value that
can generate a NaN.