fix(readme): Bring README in line with legacy support removal
This commit brings the README in line with the recent removal of support
for systems without clockid or context switch records.

This bumps up the minimum requirement for Linux to 4.3, although both
clockid and context switch records appear to be frequently backported
features.

Both --help and the manpage did not need adjustements.

This fixes #246