fix(block-io): Fix block I/O support
nr_sector and sector were switched

in the code path where we need to drop an I/O event if it arrived late,
the I/O event was never dropped, resulting in an infinite loop.

This does _not_ fix NVMe support as the cause for that issue is
unrelated to our handling of block I/O events.