feat(io-recording): Make I/O recording more generic
As a prerequisite to POSIX I/O and NVMe I/O recording, make the
MultiReader even more generic.

Neither the MultiReader, bio::Reader (renamed IoReader) or the
BioMonitor (renamed IoMonitor) contain details specific to block I/O
anymore.

To implement a new tracepoint-to-I/O recorder, one only has to implement
a Writer similiar to the bio::Writer, create the necessary definitions
in the Trace and instantiate the correct class templates