feat (I/O recording): Add POSIX I/O recording
This commit adds a "--posix-io" option to record the I/O activity of
applications that use read()/write()/open[at]() etc.

Because POSIX I/O operation are only well-ordered when viewed from the
angle of the thread executing them, POSIX I/O measurement is only
available in process monitoring mode.

This is the first lo2s component which utilizes the in-kernel virtual
machine  BPF for the recording of
events. A classical tracepoint based measurement process is not possible
for recording POSIX I/O as the filename, which is critical for
identifying what files are exactly written, is only ever passed as a
pointer to user-space memory in an application, which is not lo2s. We can normally
not read this pointer in lo2s, rendering it pretty much useless.

Using a small in-kernel BPF program, we can access the memory and copy
the name to lo2s accessible memory.

The changes introduced here result in a few new/changed major
components:

- CMake code to find and include libbpf as well as bpftool. This code
  conditionally decides wether POSIX I/O support is enabled.
- A BPF program, posix_io.bpf.c, which attachs itself to the necessary
  probe interfaces, performs the required in-kernel transformations
  of the events and writes them to a shared ring buffer.
    - The object code of this BPF program is included by the BPF toolkit
      into a "posix_io.bpf.o" object file, which also contains the
      necessary extra code to successfully load it into the kernel.
      This object files is linked into lo2s.
- A PosixMonitor, residing at the other end of BPF ring-buffer, writing
  OTF-2 events for the incoming BPF events.
- The necessary Trace suppport infrastructure to create a POSIX I/O
  writer as well as the required OTF-2 definitions.