Fix NEC sampling if threads are involved
ve_check_pid() only works for processes, not threads, so lo2s complained
that it could not find the real NEC device id for those threads.

Rewrite the code to use ve_sysfs_path_info(). This gives us the correct
sysfs path for a NEC device id, so we dont have to find it using
ve_check_pid().