feat(symbol resolution): replace libbfd with libelf/libdw
This commit lays the groundwork for replacing libbfd from binutils with
libelf and libdwarf.

Besides being 1. not documented 2. a horror-show of bad C-API design, libbfd
does not support reading DWARF debug information.[1] An upcoming feature
I am working on, however, requires DWARF.

An alternative to libbfd for the purpose of reading ELF/DWARF is
libelf/libdwarf[2]. It has been developed from the ground up with
observability in mind.

[1] Atleast I don't think that it is possible to extract DWARF via
libbfd in some meaningful manner. Some of the things it can do _feel_
like they should only be possible with DWARF debug information, but
bothering any more with this poor excuse for an API seems like a waste.

[2] https://sourceware.org/elfutils/