feat(symbol resolution): implement kallsyms support
lo2s is currently not able to resolve the symbols from instruction
pointers coming from the kernel.

I have implemented in this PR support for parsing kallsyms to alleviate
this.

/proc/kallsyms is a file which records the addresses of symbols inside
the running kernel.

One issue is that it only reports the location, not the size of symbols.
I simply assume that a symbol ends at the start of the next symbol and
the last symbol in the kernel ending at the highest address.