rework(ringbuf): Make Interface more generic
In preparation of OMPT and RocTracer support, rework the CUPTI
ringbuffer suppport to be more generic.

The new interface allows for an arbitrary amount of external event
sources to attach themselves to lo2s.

lo2s now exposes an Unix Domain Socket, per default at
/tmp/lo2s.socket.

External event sources start by creating a shared memory ring-buffer
for the recording of events.
The external event source then sends the file descriptor to that
Unix Domain Socket.

lo2s receives the file descriptor in the SocketMonitor and then opens up
a new RingbufMonitor for the received handle to shared memory.

Additionally, I have extended the calling context management.

Through out the calling context management code, the cctxs are now
tagged with the MeasurementScope instead of the Process.

This allows us to separate the MeasurementScope::sample(Process(42))
cctxs from the MeasurementScope::rb(Process(42), "cuda')
coming from CUPTI. With this, most of the cctx management code is
unified between the ring-buffers and  the sampling.

Additionally there is now a ManualFunctionResolver, which is manually
filled with the cctxs coming from the external event source.