Replace libbfd with elfutils, rework cctx tracking/function/instruction resolution
This commit accomplishes three things: Replacing libbfd, which we used for
now for symbol resolution, with elfutils, greatly simplifying the code
for the tracking the relationship between local and global calling
contexts and reworking the architecture for the resolution of function names
and instructions.

elfutils, compared to libbfd, has the advantage of having fully featured
support for DWARF information, increasing the level of information contained
in our traces if DWARF is available.

It's API is also less laden with historical cruft.

Lastly, it can make us of Debuginfod, which is a tool that can
download the debug info information for system packages on the fly if
 requested. No more installing of libfoo-dbg packages!

Because this download process can take quite some time, it is locked
behind a command line option.

For the tracking of the calling context tree, the following changes were
made:

- Use a _single_ global tree for capturing the calling contexts, with
  which is strictly treated as a tree
- Introduce a generic `CallingContext` data type that encapsulates the
  different kinds of "calling contexts" that we can have in lo2s
  (addresses, processes, threads, etc.) This is the key type of the
  tree.
    - This will come in handy later for OMPT support as we will
      introduce another type of calling context there
- Generate otf2::definition::calling_context only in the depth first
  walks of the local calling context trees, simplifying the state
  management.

Lastly, the function and instruction resolution code has been reworked:

- Split function from instruction resolvers.
- Eliminate the ProcessInfo data structure, which had a lot of bits and bobs
  of the resolution code all over it. Rather, put that code to
  places were it fits better.
- Create MemoryMap data structure with strict semantics for the
  emplacement of possibly overlapping new memory rangtes.
- Move responsibility of filling in resolvers
  from the MainMonitor to the sample::Writer. It ought to
  know better what kind of MmapEvents it has recorded and how to treat
  them.