fix(sampling): Fix issues with simultaneous thread creation
issue 155 highlights some code that makes lo2s crash due to very fast
thread enters/exits.

Most of the highlighted issues/warnings have been resolved, however lo2s
still crashes on the given code. This is due to missing OTF2 definitions
for the process the threads belong to.

In the case that there is no process, emplace a thread as its own
process. This is probably breaks some assumptions about processes (I am
pretty sure that this makes function resolution for that thread impossible,
because we track that on a process basis) in lo2s, but at least it does
not crash anymore.