Preliminary support for OpenMP recording through OMPT
The content of this commit produces traces of OpenMP program activity
which can already capture quite some information about OpenMP programs,
but at no point claim to be correct or complete or anything.

On the technical side, the functionality here is based on the ringbuf
interface, attaching lo2s to the relevant OMPT callbacks.