feat(perf): Add support for pmu-events
This adds support for pmu-events from the Linux kernel, as discussed
in issue 379.

"pmu-events" is the folder on the Linux kernel source code that contains
the list of all in-built events for the perf command. In
github.com/cvonelm/pmu-events (should probably be moved to
tud-zih-energy)
I've separated the code from the kernel and made it externally usable.

This commit adds the ability to get event names via pmu-events to lo2s.