fix(perf): use perf_event_paranoid for exclude_kernel.
Check whether exclude_kernel, or CPU_SET monitoring is available
by using the value from perf_event_paranoid. The previous code used to
check it by opening the event with exclude_kernel and checking if this
errors.

This is (on the GraceHopper for example) prone to errors unrelated to
whether exclude_kernel works or not. And due to the multi-level
error-handling nature of it all, prone to printing unhelpful messages.

On the other side, not trusting what perf_event_paranoid said on the tin
seems to be overtly paranoid. On all machines that I've tested it,
whether exclude_kernel works is solely a function of whether,
perf_event_paranoid <= 1, just as the man page of perf_event_open tells
it.