fix(tracepoints): Move from debugfs to tracefs.
In 2015[1], tracepoints were split from debugfs into their own tracefs,
mounted at `/sys/kernel/tracing`. A legacy mount of the tracefs
to the old location in the debugfs in `/sys/kernel/debug/tracing`
was maintained after that. That legacy mountpoint seems to be on the way
out, because I can not find it anymore on my personal work system.

As the `/sys/kernel/tracing` change is from 2015 and that directory
is available on any and all research and HPC system that I have access,
to, I have simply changed all mentions from debugfs -> tracefs and
changed the path everywhere to `/sys/kernel/tracing`.

[1] https://lkml.org/lkml/2015/1/26/454