tests(perf): Test for all events as defined by perf
Some perf events are "predefined events", e.g. they are defined as a
big list of #defined values (such as PERF_COUNT_HW_CPU_CYCLES) that are
independent of the selected hardware.

The perf utility uses nice human-readable names such as "cpu-cycles" for
them. lo2s wants to expose those predefined events with the same names
as perf.

This commit adds a test that parses the perf list of predefined events
and then tries to open them with lo2s. This makes sure that we stay in
sync with perf.